feat&refactor: transfer getTags to TagGroup from page.tsx and add optional catch-all segment
